+++ src/main/java/master2017/flink/AvgSpeedFine.java
@@ -2,12 +2,12 @@ package master2017.flink;
 
 import org.apache.flink.api.java.tuple.Tuple6;
 
-public class AvgSpeedFine extends Tuple6<Integer,Integer,Integer,Integer,Integer,Integer> {
+public class AvgSpeedFine extends Tuple6<Long,Long,Integer,Integer,Integer,Integer> {
     public AvgSpeedFine() {
         super();
     }
 
-    public AvgSpeedFine(int a, int b, int c, int d, int e, int f) {
+    public AvgSpeedFine(long a, long b, int c, int d, int e, int f) {
         super(a,b,c,d,e,f);
     }
 }
